requirements.txt
beautifulsoup4==4.6.1
bleach==2.1.3
Django==2.0.7
django-model-utils==3.1.2
django-widget-tweaks==1.4.2
markdown==2.6.11
Pillow==5.2.0